Project structure and reformatting
