+++ src/App.vue
@@ -5,12 +5,14 @@
       <div class="navbar-nav mr-auto">
         <li class="nav-item">
           <router-link to="/home" class="nav-link">
-            <font-awesome-icon icon="home" />Home
+            <font-awesome-icon icon="home"/>
+            Home
           </router-link>
         </li>
         <li class="nav-item">
           <router-link to="/cars" class="nav-link">
-            <font-awesome-icon icon="home" />Cars
+            <font-awesome-icon icon="home"/>
+            Cars
           </router-link>
         </li>
         <li v-if="showManagerBoard" class="nav-item">
@@ -24,12 +26,12 @@
       <div v-if="!currentUser" class="navbar-nav ml-auto">
         <li class="nav-item">
           <router-link to="/register" class="nav-link">
-           <i class="fa-solid fa-user-plus"></i> Become a driver
+            <i class="fa-solid fa-user-plus"></i> Become a driver
           </router-link>
         </li>
         <li class="nav-item">
           <router-link to="/login" class="nav-link">
-             <i class="fa-solid fa-user-lock"></i> Login
+            <i class="fa-solid fa-user-lock"></i> Login
           </router-link>
         </li>
       </div>
@@ -50,7 +52,7 @@
     </nav>
 
     <div class="container">
-      <router-view />
+      <router-view/>
     </div>
   </div>
 </template>
