+++ src/main.js
@@ -1,4 +1,4 @@
-import { createApp } from "vue";
+import {createApp} from "vue";
 import App from "./App.vue";
 import router from "./router";
 import store from "./store";
@@ -6,6 +6,6 @@ import "bootstrap";
 import "bootstrap/dist/css/bootstrap.min.css";
 
 createApp(App)
-  .use(router)
-  .use(store)
-  .mount("#app");
+    .use(router)
+    .use(store)
+    .mount("#app");
