+++ src/services/auth.service.js
@@ -3,32 +3,32 @@ import axios from 'axios';
 const API_URL = 'http://localhost:8081/api/auth/';
 
 class AuthService {
-  login(user) {
-    return axios
-      .post(API_URL + 'signin', {
-        username: user.username,
-        password: user.password
-      })
-      .then(response => {
-        if (response.data.accessToken) {
-          localStorage.setItem('user', JSON.stringify(response.data));
-        }
+    login(user) {
+        return axios
+            .post(API_URL + 'signin', {
+                username: user.username,
+                password: user.password
+            })
+            .then(response => {
+                if (response.data.accessToken) {
+                    localStorage.setItem('user', JSON.stringify(response.data));
+                }
 
-        return response.data;
-      });
-  }
+                return response.data;
+            });
+    }
 
-  logout() {
-    localStorage.removeItem('user');
-  }
+    logout() {
+        localStorage.removeItem('user');
+    }
 
-  register(user) {
-    return axios.post(API_URL + 'signup', {
-      username: user.username,
-      email: user.email,
-      password: user.password
-    });
-  }
+    register(user) {
+        return axios.post(API_URL + 'signup', {
+            username: user.username,
+            email: user.email,
+            password: user.password
+        });
+    }
 }
 
 export default new AuthService();
