+++ src/store/index.js
@@ -1,10 +1,10 @@
-import { createStore } from "vuex";
-import { auth } from "./auth.module";
+import {createStore} from "vuex";
+import {auth} from "./auth.module";
 
 const store = createStore({
-  modules: {
-    auth,
-  },
+    modules: {
+        auth,
+    },
 });
 
 export default store;
