+++ src/views/BoardManager.vue
@@ -18,17 +18,17 @@ export default {
   },
   mounted() {
     UserService.getManagerBoard().then(
-      (response) => {
-        this.content = response.data;
-      },
-      (error) => {
-        this.content =
-          (error.response &&
-            error.response.data &&
-            error.response.data.message) ||
-          error.message ||
-          error.toString();
-      }
+        (response) => {
+          this.content = response.data;
+        },
+        (error) => {
+          this.content =
+              (error.response &&
+                  error.response.data &&
+                  error.response.data.message) ||
+              error.message ||
+              error.toString();
+        }
     );
   },
 };
