+++ tests/unit/userhome.spec.js
@@ -1,12 +1,12 @@
-import { shallowMount } from '@vue/test-utils'
+import {shallowMount} from '@vue/test-utils'
 import UserHome from '@/components/UserHome.vue'
 
 describe('UserHome.vue', () => {
-  it('renders Lorem ipsum when passed', () => {
-    const msg = 'Lorem ipsum'
-    const wrapper = shallowMount(UserHome, {
-      props: { msg }
+    it('renders Lorem ipsum when passed', () => {
+        const msg = 'Lorem ipsum'
+        const wrapper = shallowMount(UserHome, {
+            props: {msg}
+        })
+        expect(wrapper.text()).toMatch(msg)
     })
-    expect(wrapper.text()).toMatch(msg)
-  })
 })
