+++ vue.config.js
@@ -1,7 +1,7 @@
-const { defineConfig } = require('@vue/cli-service')
+const {defineConfig} = require('@vue/cli-service')
 module.exports = defineConfig({
-  devServer:{
-    port:8081
-  },
-  transpileDependencies: true
+    devServer: {
+        port: 8081
+    },
+    transpileDependencies: true
 })
